% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api_link}
\alias{api_link}
\title{Create link for API}
\usage{
api_link(...)
}
\arguments{
\item{...}{Arguments to specify the path to the API endpoint}
}
\value{
Formatted link to API endpoint
}
\description{
Create link for API
}
\keyword{internal}
