% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api_page_size}
\alias{api_page_size}
\title{Extract the page size from API metadata}
\usage{
api_page_size(messages)
}
\arguments{
\item{messages}{API metadata messages data frame}
}
\value{
Numeric page size
}
\description{
Extract the page size from API metadata
}
\keyword{internal}
