% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api_record_count}
\alias{api_record_count}
\title{Extract the total number of records from API metadata}
\usage{
api_record_count(messages)
}
\arguments{
\item{messages}{API metadata messages data frame}
}
\value{
Numeric record count
}
\description{
Extract the total number of records from API metadata
}
\keyword{internal}
