% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api_to_df}
\alias{api_to_df}
\title{Convert API data to data frame}
\usage{
api_to_df(url)
}
\arguments{
\item{url}{API endpoint from which to extract and format data}
}
\value{
Raw API data in a dataframe
}
\description{
Convert API data to data frame
}
\examples{
\dontrun{
df <- api_to_df(api_link("medrxiv", "10.1101/2020.02.25.20021568"))
}
}
\keyword{internal}
