% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_api_df}
\alias{clean_api_df}
\title{Helper script to clean data from API to make it compatible with mx_search()}
\usage{
clean_api_df(df)
}
\arguments{
\item{df}{Raw dataframe from API}
}
\value{
Cleaned dataframe
}
\description{
Helper script to clean data from API to make it compatible with mx_search()
}
\keyword{internal}
