% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_syntax.R
\name{fix_near}
\alias{fix_near}
\title{Replace user-friendly 'NEAR' operator with appropriate regex syntax}
\usage{
fix_near(x)
}
\arguments{
\item{x}{Search query to be reformatted}
}
\description{
Replace user-friendly 'NEAR' operator with appropriate regex syntax
}
\keyword{internal}
