% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_snapshot.R
\name{inform_snapshot_date}
\alias{inform_snapshot_date}
\title{Report the latest record date in a snapshot}
\usage{
inform_snapshot_date(data)
}
\arguments{
\item{data}{Snapshot data frame}
}
\value{
Invisibly returns the latest snapshot date
}
\description{
Report the latest record date in a snapshot
}
\keyword{internal}
