% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{internet_check}
\alias{internet_check}
\title{Checks whether the user has internet, and returns a helpful message it not.}
\usage{
internet_check()
}
\value{
Informative error if not connected to the internet
}
\description{
Checks whether the user has internet, and returns a helpful message it not.
}
\keyword{Internal}
