% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medrxivr-package.R
\docType{package}
\name{medrxivr-package}
\alias{medrxivr}
\alias{medrxivr-package}
\title{medrxivr: Accessing medRxiv and bioRxiv preprint data from R}
\description{
The medrxivr package enables users to access data on preprints in the medRxiv
and bioRxiv preprints repositories, both of which are run by the Cold Spring
Harbour Laboratory.It also provides functions to search the preprint data,
export it to a .bib file, and download the PDFs associated with specified
records.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/medrxivr/}
  \item \url{https://github.com/ropensci/medrxivr}
  \item Report bugs at \url{https://github.com/ropensci/medrxivr/issues}
}

}
\author{
\strong{Maintainer}: Yaoxiang Li \email{liyaoxiang@outlook.com} (\href{https://orcid.org/0000-0001-9200-1016}{ORCID})

Authors:
\itemize{
  \item Luke McGuinness
  \item Lena Schmidt
}

Other contributors:
\itemize{
  \item Tuija Sonkkila [reviewer]
  \item Najko Jahn [reviewer]
}

}
\keyword{internal}
