% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_search.R
\name{mx_reporter}
\alias{mx_reporter}
\title{Search and print output for individual search items}
\usage{
mx_reporter(mx_data, query, fields, NOT, deduplicate)
}
\arguments{
\item{mx_data}{The mx_dataset filtered for the date limits}

\item{query}{Character string, vector or list}

\item{fields}{Fields of the database to search - default is Title, Abstract,
Authors, Category, and DOI.}

\item{NOT}{Vector of regular expressions to exclude from the search. Default
is NULL.}

\item{deduplicate}{Logical. Only return the most recent version of a record.
Default is TRUE.}
}
\description{
Search and print output for individual search items
}
\seealso{
Other main: 
\code{\link{mx_search}()},
\code{\link{print_full_results}()},
\code{\link{run_search}()}
}
\concept{main}
