% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_search.R
\name{print_full_results}
\alias{print_full_results}
\title{Search for terms in the dataset}
\usage{
print_full_results(num_results, deduplicate)
}
\arguments{
\item{num_results}{number of searched terms returned}

\item{deduplicate}{Logical. Only return the most recent version of a record.
Default is TRUE.}
}
\description{
Search for terms in the dataset
}
\seealso{
Other main: 
\code{\link{.mx_safe_fix_caps}()},
\code{\link{mx_reporter}()},
\code{\link{run_search}()}
}
\concept{main}
