% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_search.R
\name{run_search}
\alias{run_search}
\title{Search for terms in the dataset}
\usage{
run_search(mx_data, query, fields, deduplicate, NOT = "")
}
\arguments{
\item{mx_data}{The mx_dataset filtered for the date limits}

\item{query}{Character string, vector or list}

\item{fields}{Fields of the database to search - default is Title, Abstract,
Authors, Category, and DOI.}

\item{deduplicate}{Logical. Only return the most recent version of a record.
Default is TRUE.}

\item{NOT}{Vector of regular expressions to exclude from the search. Default
is NULL.}
}
\description{
Search for terms in the dataset
}
\seealso{
Other main: 
\code{\link{.mx_safe_fix_caps}()},
\code{\link{mx_reporter}()},
\code{\link{print_full_results}()}
}
\concept{main}
