% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{skip_if_api_message}
\alias{skip_if_api_message}
\title{Skips API tests if API isn't working correctly}
\usage{
skip_if_api_message()
}
\description{
Skips API tests if API isn't working correctly
}
\examples{
\dontrun{
skip_if_api_message()
}
}
\keyword{internal}
