% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api_link}
\alias{api_link}
\title{Create link for API}
\usage{
api_link(...)
}
\arguments{
\item{...}{Arguments to specify the path to the API endpoint}
}
\value{
Formatted link to API endpoint
}
\description{
Create link for API
}
\examples{
\dontrun{
link <- api_link("details", "2020-01-01", "2020-01-31", "0")
link <- api_link("details", "10.1101/2020.02.25.20021568")
}

}
\keyword{internal}
