% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_api.R
\name{mx_api_content}
\alias{mx_api_content}
\title{Access Cold Springs Harbour API, which serves both medRxiv and bioRxiv}
\usage{
mx_api_content(
  from_date = "2013-01-01",
  to_date = as.character(Sys.Date()),
  clean = TRUE,
  server = "medrxiv",
  include.info = FALSE
)
}
\arguments{
\item{from_date}{Earliest date of interest. Defaults to 1st June 2019
(earliest medRxiv record was posted on 25th June 2019).}

\item{to_date}{Latest date of interest. Defaults to current date.}

\item{clean}{Logical, indicating whether to clean the data returned for use
with other mx_* functions. Default is TRUE.}

\item{server}{Specify the server you wish to use: "medrxiv" (default) or
"biorxiv"}

\item{include.info}{Logical, indicating whether to include variables
containing information returned by the API (e.g. cursor number, total count
of papers, etc). Default is FALSE.}
}
\value{
Dataframe with 1 record per row
}
\description{
Access Cold Springs Harbour API, which serves both medRxiv and bioRxiv
}
\examples{
\dontrun{
mx_data <- mx_api_content()
}
}
\seealso{
Other data-source: 
\code{\link{mx_api_doi}()},
\code{\link{mx_snapshot}()}
}
\concept{data-source}
