% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_api.R
\name{mx_api_doi}
\alias{mx_api_doi}
\title{Access medRxiv API, based on single DOI}
\usage{
mx_api_doi(doi, server = "medrxiv", clean = TRUE)
}
\arguments{
\item{doi}{Digital object identifier of the preprint you wish to retrieve
data on.}

\item{server}{Specify the server you wish to use: "medrxiv" (default) or
"biorxiv"}

\item{clean}{Logical, indicating whether to clean the data returned for use
with other mx_* functions.}
}
\value{
Dataframe containing details on the preprint identified by the DOI.
}
\description{
Access medRxiv API, based on single DOI
}
\examples{
\dontrun{
mx_data <- mx_api_doi("10.1101/2020.02.25.20021568")
}
}
\seealso{
Other api: 
\code{\link{mx_api_content}()}
}
\concept{api}
