% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_crosscheck.R
\name{mx_crosscheck}
\alias{mx_crosscheck}
\title{Cross-check the medRxiv snapshot against the Cold Spring Harbour Laboratory API}
\usage{
mx_crosscheck()
}
\description{
Provides information on how up-to-date the maintained medRxiv
  snapshot is by checking whether there have been any records added
  to/updated in the medRxiv repository since the last snapshot was taken.
}
\examples{
\dontrun{
mx_crosscheck()
}
}
\concept{helper}
