% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_crosscheck.R
\name{mx_crosscheck}
\alias{mx_crosscheck}
\title{Live check against medRxiv}
\usage{
mx_crosscheck()
}
\description{
Cross-check whether the dataset you are using is current
}
\examples{
\dontrun{
mx_crosscheck()
}
}
\concept{helper}
