% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_download.R
\name{mx_download}
\alias{mx_download}
\title{Download PDF's}
\usage{
mx_download(mx_results, directory, create = TRUE, print_update = 10)
}
\arguments{
\item{mx_results}{Vector containing the links to the medRxiv PDFs}

\item{directory}{The location you want to download the PDF's to}

\item{create}{TRUE or FALSE. If TRUE, creates the directory if it doesn't exist}

\item{print_update}{How frequently to print an update}
}
\description{
Download PDF's of all the papers in your search results
}
\examples{
\dontrun{
mx_results <- mx_search("ecology",limit=20)
mx_download(mx_results,"~/medrxivPDF")
}
}
