% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_export.R
\name{mx_export}
\alias{mx_export}
\title{Export results of medRxiv search to a .bib file}
\usage{
mx_export(data, file = "medrxiv_export.bib")
}
\arguments{
\item{data}{Dataframe returned by mx_search()}

\item{file}{File location to save to. Must have the .bib file extension}
}
\value{
Exports a formatted .BIB file, for import into a reference manager
}
\description{
Export results of medRxiv search to a .bib file
}
\examples{
\dontrun{
mx_results <- mx_search(mx_snapshot(), query = "brain")
mx_export(mx_results, "references.bib")
}

}
