% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_info.R
\name{mx_info}
\alias{mx_info}
\title{Provide information on the medRxiv snapshot used to perform the search}
\usage{
mx_info(commit = "master")
}
\arguments{
\item{commit}{Commit hash for the snapshot, taken from
https://github.com/mcguinlu/medrxivr-data. Defaults to "master", which will
return info on the most recent snapshot.}
}
\value{
Message with snapshot details
}
\description{
Provide information on the medRxiv snapshot used to perform the search
}
\keyword{internal}
