% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_search.R
\name{mx_search}
\alias{mx_search}
\title{Search medRxiv}
\usage{
mx_search(query, limit = 10, NOT = NULL)
}
\arguments{
\item{query}{Vector containing the links to the medRxiv PDFs}

\item{limit}{The number of results to return}
}
\description{
Search medRxiv using a string
}
\examples{
\dontrun{
mx_results <- mx_search("dementia",limit=20)
}
}
