% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_search.R
\name{mx_search}
\alias{mx_search}
\title{Search medRxiv}
\usage{
mx_search(query, limit = 10, NOT = NULL, deduplicate = FALSE)
}
\arguments{
\item{query}{Character string, vector or list}

\item{limit}{The number of results to return}

\item{NOT}{Vector of regular expressions to exclude from the search. Default is NULL.}

\item{deduplicate}{Logical. Only return the most recent version of a record. Default is TRUE}
}
\description{
Search medRxiv using a string
}
\examples{
\dontrun{
mx_results <- mx_search("dementia",limit=20)
}
}
