% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_search.R
\name{mx_search}
\alias{mx_search}
\title{Search preprint data}
\usage{
mx_search(
  data = NULL,
  query = NULL,
  fields = c("title", "abstract", "authors", "category", "doi"),
  from_date = NULL,
  to_date = NULL,
  NOT = "",
  deduplicate = TRUE
)
}
\arguments{
\item{data}{The preprint dataset that is to be searched, created either using
mx_api_content() or mx_snapshot()}

\item{query}{Character string, vector or list}

\item{fields}{Fields of the database to search - default is Title, Abstract,
Authors, Category, and DOI.}

\item{from_date}{Defines earliest date of interest. Written in the format
"YYYY-MM-DD". Note, records published on the date specified will also be
returned.}

\item{to_date}{Defines latest date of interest. Written in the format
"YYYY-MM-DD". Note, records published on the date specified will also be
returned.}

\item{NOT}{Vector of regular expressions to exclude from the search. Default
is NULL.}

\item{deduplicate}{Logical. Only return the most recent version of a record.
Default is TRUE.}
}
\description{
Search preprint data
}
\examples{
\dontrun{
# Using data from the API
mx_results <- mx_api_content() \%>\%
  mx_search(query = "dementia")

# Using the daily snapshot
mx_results <- mx_snapshot() \%>\%
  mx_search(query = "dementia")
}
}
\concept{main}
