% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_snapshot.R
\name{mx_snapshot}
\alias{mx_snapshot}
\title{Returns formatted snapshot of the medRxiv database for use with mx_search()}
\usage{
mx_snapshot(commit = "master")
}
\arguments{
\item{commit}{Commit hash for the snapshot, taken from
https://github.com/mcguinlu/medrxivr-data. Allows for reproducible
searching by specifying the exact snapshot used to perform the searches.
Defaults to "master", which will return the most recent snapshot.}
}
\value{
Formatted dataframe
}
\description{
Returns formatted snapshot of the medRxiv database for use with mx_search()
}
\examples{
\dontrun{
mx_data <- mx_snapshot()
}

}
