% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_snapshot.R
\name{mx_snapshot}
\alias{mx_snapshot}
\title{Access a static snapshot of the medRxiv repository}
\usage{
mx_snapshot(commit = "main")
}
\arguments{
\item{commit}{Commit hash or branch name for the snapshot, taken from
https://github.com/yaoxiangli/medrxivr-data. Allows for reproducible searching
by specifying the exact snapshot used to perform the searches. Defaults to
"main", which will return the most recent snapshot from the main branch.}
}
\value{
A formatted dataframe containing the combined data from the snapshot
  parts, with reconstructed `link_page` and `link_pdf` columns.
}
\description{
[Available for medRxiv only] This function allows users to import
  a maintained static snapshot of the medRxiv repository, instead of downloading
  a copy from the API, which can become unavailable during peak usage times.
  The function dynamically retrieves multiple snapshot parts from the specified
  repository and combines them into a single dataframe.
}
\examples{
\donttest{
mx_data <- mx_snapshot()
mx_data_specific <- mx_snapshot(commit = "specific_commit_hash")
}

}
\seealso{
Other data-source: 
\code{\link{mx_api_content}()},
\code{\link{mx_api_doi}()}
}
\concept{data-source}
