% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_snapshot.R
\name{mx_snapshot}
\alias{mx_snapshot}
\title{Access a static snapshot of the medRxiv repository}
\usage{
mx_snapshot(commit = "master")
}
\arguments{
\item{commit}{Commit hash for the snapshot, taken from
https://github.com/mcguinlu/medrxivr-data. Allows for reproducible
searching by specifying the exact snapshot used to perform the searches.
Defaults to "master", which will return the most recent snapshot.}
}
\value{
Formatted dataframe
}
\description{
[Available for medRxiv only] Rather than downloading a copy of
  the medRxiv database from the API, which can become unavailable at peak
  usage times, this allows users to import a maintained static snapshot of
  the medRxiv repository.
}
\examples{
\donttest{
mx_data <- mx_snapshot()
}

}
\seealso{
Other data-source: 
\code{\link{mx_api_content}()},
\code{\link{mx_api_doi}()}
}
\concept{data-source}
